refactor(mcp): extract resource handlers + proxy branches (lab-kvji.24.1.3)
list_resources/read_resource -> inherent *_impl in handlers_resources.rs;
read_resource's three proxy branches (gateway-synthetic, upstream, subject-scoped)
-> inherent methods in resource_proxy.rs. read_resource_impl keeps the prefix-dispatch
skeleton + local lab://catalog / lab://<svc>/actions branch.

Seam discipline (Revision 2 #4): subject/uri/pool/config/oauth_subject threaded from
caller guards; three-branch ordering and per-branch logging side-effects byte-identical
(no circuit-breaker record_* in resource path). ServerHandler keeps one-line delegators.

handlers_resources.rs 235, resource_proxy.rs 354 LOC. server.rs 2847 -> 2382. 1423 tests pass.
