+++ crates/lab/src/mcp.rs
@@ -9,11 +9,13 @@ pub mod elicitation;
 pub mod envelope;
 pub mod error;
 pub mod handlers_prompts;
+pub mod handlers_resources;
 pub mod logging;
 pub mod meta;
 pub mod peers;
 pub mod prompts;
 pub mod registry;
+pub mod resource_proxy;
 pub mod resources;
 pub mod result_format;
 pub mod server;
