+++ crates/lab/src/mcp/handlers_resources.rs
@@ -1,0 +1,235 @@
+//! Resource handler bodies (`list_resources`, `read_resource`).
+//!
+//! Extracted from `server.rs` (bead `lab-kvji.24.1.3`) as inherent
+//! `impl LabMcpServer` methods. The `ServerHandler` trait impl in
+//! `server.rs` keeps one-line delegators.
+//!
+//! `read_resource_impl` keeps the prefix-dispatch skeleton + the local
+//! `lab://catalog` / `lab://<svc>/actions` branch; the three proxy
+//! branches live in `resource_proxy.rs` and are reached via the same
+//! guard ordering as the original (gateway → upstream → subject-scoped).
+//!
+//! No behavior change — relocation only.
+
+use std::time::Instant;
+
+use rmcp::ErrorData;
+use rmcp::RoleServer;
+use rmcp::model::{
+    AnnotateAble, ListResourcesResult, LoggingLevel, PaginatedRequestParams, RawResource,
+    ReadResourceRequestParams, ReadResourceResult, ResourceContents,
+};
+use rmcp::service::RequestContext;
+
+use crate::mcp::context::{auth_context_from_extensions, oauth_upstream_subject_for_request};
+use crate::mcp::logging::DispatchLogOutcome;
+use crate::mcp::server::LabMcpServer;
+
+impl LabMcpServer {
+    pub(crate) async fn list_resources_impl(
+        &self,
+        _request: Option<PaginatedRequestParams>,
+        context: RequestContext<RoleServer>,
+    ) -> Result<ListResourcesResult, ErrorData> {
+        let start = Instant::now();
+        let subject = self.request_subject_log_tag(&context);
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "list_resources",
+            subject,
+            "dispatch start"
+        );
+        let mut resources = vec![
+            RawResource::new("lab://catalog", "catalog")
+                .with_description("Full discovery document for all services")
+                .with_mime_type("application/json")
+                .no_annotation(),
+        ];
+
+        for svc in self.registry.services() {
+            if self.service_visible_on_mcp(svc.name).await {
+                let uri = format!("lab://{}/actions", svc.name);
+                let name = format!("{}/actions", svc.name);
+                resources.push(
+                    RawResource::new(uri, name)
+                        .with_description(format!("Action list for {}", svc.name))
+                        .with_mime_type("application/json")
+                        .no_annotation(),
+                );
+            }
+        }
+
+        if let Some(pool) = self.current_upstream_pool().await {
+            resources.extend(pool.gateway_synthetic_resources().await);
+            resources.extend(pool.list_upstream_resources().await);
+            let auth = auth_context_from_extensions(&context.extensions);
+            if let Some(oauth_subject) =
+                oauth_upstream_subject_for_request(auth, self.request_subject(&context))
+            {
+                let configs = self.oauth_upstream_configs().await;
+                resources.extend(
+                    pool.subject_scoped_resources(&configs, oauth_subject.as_ref())
+                        .await,
+                );
+            }
+        }
+
+        let elapsed_ms = start.elapsed().as_millis();
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "list_resources",
+            subject,
+            elapsed_ms,
+            "resource list ok"
+        );
+        self.emit_dispatch_notification(
+            &context,
+            "lab",
+            "list_resources",
+            elapsed_ms,
+            DispatchLogOutcome::Success,
+        )
+        .await;
+
+        Ok(ListResourcesResult::with_all_items(resources))
+    }
+
+    pub(crate) async fn read_resource_impl(
+        &self,
+        request: ReadResourceRequestParams,
+        context: RequestContext<RoleServer>,
+    ) -> Result<ReadResourceResult, ErrorData> {
+        let start = Instant::now();
+        let subject = self.request_subject_log_tag(&context);
+        let uri = &request.uri;
+        tracing::info!(
+            surface = "mcp",
+            service = "labby",
+            action = "read_resource",
+            subject,
+            resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
+            "dispatch start"
+        );
+
+        // Branch 1: gateway-synthetic resources.
+        if uri.starts_with("lab://gateway/") {
+            return self
+                .read_gateway_resource_impl(uri, &subject, start, &context)
+                .await;
+        }
+
+        // Branch 2: raw upstream resource proxy.
+        if let Some(pool) = self.current_upstream_pool().await
+            && uri.starts_with("lab://upstream/")
+        {
+            return self
+                .read_upstream_resource_impl(&pool, uri, &subject, start, &context)
+                .await;
+        }
+
+        // Branch 3: subject-scoped upstream resource proxy.
+        let auth = auth_context_from_extensions(&context.extensions);
+        if let Some(oauth_subject) =
+            oauth_upstream_subject_for_request(auth, self.request_subject(&context))
+            && let Some(pool) = self.current_upstream_pool().await
+            && let Some(upstream_name) = uri
+                .strip_prefix("lab://upstream/")
+                .and_then(|rest| rest.split('/').next())
+            && let Some(config) = self.oauth_upstream_config(upstream_name).await
+        {
+            return self
+                .read_subject_scoped_resource_impl(
+                    &pool,
+                    &config,
+                    oauth_subject.as_ref(),
+                    uri,
+                    &subject,
+                    start,
+                    &context,
+                )
+                .await;
+        }
+
+        // Local branch: lab://catalog + lab://<svc>/actions.
+        let json = if uri == "lab://catalog" {
+            self.catalog_json().await
+        } else if let Some(service) = uri
+            .strip_prefix("lab://")
+            .and_then(|value| value.strip_suffix("/actions"))
+        {
+            self.service_actions_json(service).await
+        } else {
+            return Err(ErrorData::resource_not_found(
+                format!("unknown resource: {uri}"),
+                None,
+            ));
+        };
+
+        match json {
+            Ok(value) => {
+                let text = match serde_json::to_string_pretty(&value) {
+                    Ok(t) => t,
+                    Err(e) => {
+                        tracing::error!(
+                            surface = "mcp",
+                            service = "labby",
+                            action = "read_resource",
+                            subject,
+                            error = %e,
+                            "failed to serialize resource"
+                        );
+                        return Err(ErrorData::internal_error(
+                            format!("failed to serialize resource: {e}"),
+                            None,
+                        ));
+                    }
+                };
+                let elapsed_ms = start.elapsed().as_millis();
+                tracing::info!(
+                    surface = "mcp",
+                    service = "labby",
+                    action = "read_resource",
+                    subject,
+                    elapsed_ms,
+                    "resource read ok"
+                );
+                self.emit_dispatch_notification(
+                    &context,
+                    "lab",
+                    "read_resource",
+                    elapsed_ms,
+                    DispatchLogOutcome::Success,
+                )
+                .await;
+                Ok(ReadResourceResult::new(vec![
+                    ResourceContents::text(text, uri.clone()).with_mime_type("application/json"),
+                ]))
+            }
+            Err(e) => {
+                let elapsed_ms = start.elapsed().as_millis();
+                tracing::error!(
+                    surface = "mcp",
+                    service = "labby",
+                    action = "read_resource",
+                    elapsed_ms,
+                    kind = "internal_error",
+                    "resource read failed"
+                );
+                self.emit_dispatch_notification(
+                    &context,
+                    "lab",
+                    "read_resource",
+                    elapsed_ms,
+                    DispatchLogOutcome::Failure {
+                        level: LoggingLevel::Error,
+                        kind: "internal_error",
+                    },
+                )
+                .await;
+                Err(ErrorData::internal_error(e.to_string(), None))
+            }
+        }
+    }
+}
