+++ crates/lab/src/mcp/server.rs
@@ -9,11 +9,10 @@ use std::time::Instant;
 
 use axum::http;
 use rmcp::model::{
-    AnnotateAble, CallToolRequestParams, CallToolResult, CompleteRequestParams, CompleteResult,
-    Content, GetPromptRequestParams, GetPromptResult, ListPromptsResult, ListResourcesResult,
-    ListToolsResult, LoggingLevel, PaginatedRequestParams, RawResource, ReadResourceRequestParams,
-    ReadResourceResult, ResourceContents, ServerCapabilities, ServerInfo, SetLevelRequestParams,
-    Tool,
+    CallToolRequestParams, CallToolResult, CompleteRequestParams, CompleteResult, Content,
+    GetPromptRequestParams, GetPromptResult, ListPromptsResult, ListResourcesResult,
+    ListToolsResult, LoggingLevel, PaginatedRequestParams, ReadResourceRequestParams,
+    ReadResourceResult, ServerCapabilities, ServerInfo, SetLevelRequestParams, Tool,
 };
 use rmcp::service::{NotificationContext, Peer, RequestContext};
 use rmcp::{ErrorData, RoleServer, ServerHandler};
@@ -310,72 +309,10 @@ impl ServerHandler for LabMcpServer {
 
     async fn list_resources(
         &self,
-        _request: Option<PaginatedRequestParams>,
+        request: Option<PaginatedRequestParams>,
         context: RequestContext<RoleServer>,
     ) -> Result<ListResourcesResult, ErrorData> {
-        let start = Instant::now();
-        let subject = self.request_subject_log_tag(&context);
-        tracing::info!(
-            surface = "mcp",
-            service = "labby",
-            action = "list_resources",
-            subject,
-            "dispatch start"
-        );
-        let mut resources = vec![
-            RawResource::new("lab://catalog", "catalog")
-                .with_description("Full discovery document for all services")
-                .with_mime_type("application/json")
-                .no_annotation(),
-        ];
-
-        for svc in self.registry.services() {
-            if self.service_visible_on_mcp(svc.name).await {
-                let uri = format!("lab://{}/actions", svc.name);
-                let name = format!("{}/actions", svc.name);
-                resources.push(
-                    RawResource::new(uri, name)
-                        .with_description(format!("Action list for {}", svc.name))
-                        .with_mime_type("application/json")
-                        .no_annotation(),
-                );
-            }
-        }
-
-        if let Some(pool) = self.current_upstream_pool().await {
-            resources.extend(pool.gateway_synthetic_resources().await);
-            resources.extend(pool.list_upstream_resources().await);
-            let auth = auth_context_from_extensions(&context.extensions);
-            if let Some(oauth_subject) =
-                oauth_upstream_subject_for_request(auth, self.request_subject(&context))
-            {
-                let configs = self.oauth_upstream_configs().await;
-                resources.extend(
-                    pool.subject_scoped_resources(&configs, oauth_subject.as_ref())
-                        .await,
-                );
-            }
-        }
-
-        let elapsed_ms = start.elapsed().as_millis();
-        tracing::info!(
-            surface = "mcp",
-            service = "labby",
-            action = "list_resources",
-            subject,
-            elapsed_ms,
-            "resource list ok"
-        );
-        self.emit_dispatch_notification(
-            &context,
-            "lab",
-            "list_resources",
-            elapsed_ms,
-            DispatchLogOutcome::Success,
-        )
-        .await;
-
-        Ok(ListResourcesResult::with_all_items(resources))
+        self.list_resources_impl(request, context).await
     }
 
     async fn read_resource(
@@ -383,409 +320,7 @@ impl ServerHandler for LabMcpServer {
         request: ReadResourceRequestParams,
         context: RequestContext<RoleServer>,
     ) -> Result<ReadResourceResult, ErrorData> {
-        let start = Instant::now();
-        let subject = self.request_subject_log_tag(&context);
-        let uri = &request.uri;
-        tracing::info!(
-            surface = "mcp",
-            service = "labby",
-            action = "read_resource",
-            subject,
-            resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-            "dispatch start"
-        );
-
-        if uri.starts_with("lab://gateway/") {
-            tracing::info!(
-                surface = "mcp",
-                service = "labby",
-                action = "read_resource",
-                subject,
-                resource_uri =
-                    crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                route = "gateway",
-                "dispatch route selected"
-            );
-            let Some(pool) = self.current_upstream_pool().await else {
-                let elapsed_ms = start.elapsed().as_millis();
-                tracing::warn!(
-                    surface = "mcp",
-                    service = "labby",
-                    action = "read_resource",
-                    subject,
-                    resource_uri =
-                        crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                    route = "gateway",
-                    elapsed_ms,
-                    kind = "unavailable",
-                    "upstream pool not configured"
-                );
-                self.emit_dispatch_notification(
-                    &context,
-                    "lab",
-                    "read_resource",
-                    elapsed_ms,
-                    DispatchLogOutcome::Failure {
-                        level: LoggingLevel::Warning,
-                        kind: "unavailable",
-                    },
-                )
-                .await;
-                return Err(ErrorData::resource_not_found(
-                    "upstream pool not configured".to_string(),
-                    None,
-                ));
-            };
-
-            let json = if uri == "lab://gateway/servers" {
-                Some(pool.gateway_servers_doc().await)
-            } else if let Some(name) = uri
-                .strip_prefix("lab://gateway/")
-                .and_then(|rest| rest.strip_suffix("/schema"))
-                .filter(|name| !name.is_empty() && !name.contains('/'))
-            {
-                pool.gateway_server_schema(name).await
-            } else {
-                None
-            };
-
-            let elapsed_ms = start.elapsed().as_millis();
-            return match json {
-                Some(value) => {
-                    let text = match serde_json::to_string_pretty(&value) {
-                        Ok(t) => t,
-                        Err(e) => {
-                            tracing::error!(
-                                surface = "mcp",
-                                service = "labby",
-                                action = "read_resource",
-                                resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                                error = %e,
-                                "failed to serialize synthetic gateway resource"
-                            );
-                            return Err(ErrorData::internal_error(
-                                format!("failed to serialize resource: {e}"),
-                                None,
-                            ));
-                        }
-                    };
-                    tracing::info!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        subject,
-                        resource_uri =
-                            crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        route = "gateway",
-                        elapsed_ms,
-                        "synthetic resource ok"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Success,
-                    )
-                    .await;
-                    Ok(ReadResourceResult::new(vec![
-                        ResourceContents::text(text, uri.clone())
-                            .with_mime_type("application/json"),
-                    ]))
-                }
-                None => {
-                    tracing::warn!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        subject,
-                        resource_uri =
-                            crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        route = "gateway",
-                        elapsed_ms,
-                        kind = "not_found",
-                        "synthetic resource not found"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Failure {
-                            level: LoggingLevel::Warning,
-                            kind: "not_found",
-                        },
-                    )
-                    .await;
-                    Err(ErrorData::resource_not_found(
-                        format!("unknown resource: {uri}"),
-                        None,
-                    ))
-                }
-            };
-        }
-
-        if let Some(pool) = self.current_upstream_pool().await
-            && uri.starts_with("lab://upstream/")
-        {
-            tracing::info!(
-                surface = "mcp",
-                service = "labby",
-                action = "read_resource",
-                resource_uri =
-                    crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                route = "upstream",
-                "dispatch route selected"
-            );
-            let outcome = match pool.read_upstream_resource(uri).await {
-                Some(Ok(result)) => {
-                    let elapsed_ms = start.elapsed().as_millis();
-                    let upstream = uri
-                        .strip_prefix("lab://upstream/")
-                        .and_then(|rest| rest.split('/').next())
-                        .unwrap_or("unknown");
-                    tracing::info!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        subject,
-                        upstream,
-                        resource_uri =
-                            crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        elapsed_ms,
-                        "resource proxy ok"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Success,
-                    )
-                    .await;
-                    Ok(result)
-                }
-                Some(Err(message)) => {
-                    let elapsed_ms = start.elapsed().as_millis();
-                    let upstream = uri
-                        .strip_prefix("lab://upstream/")
-                        .and_then(|rest| rest.split('/').next())
-                        .unwrap_or("unknown");
-                    tracing::warn!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        upstream,
-                        resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        elapsed_ms,
-                        kind = "internal_error",
-                        error = %message,
-                        "resource proxy failed"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Failure {
-                            level: LoggingLevel::Error,
-                            kind: "internal_error",
-                        },
-                    )
-                    .await;
-                    Err(ErrorData::internal_error(message, None))
-                }
-                None => {
-                    let elapsed_ms = start.elapsed().as_millis();
-                    let upstream = uri
-                        .strip_prefix("lab://upstream/")
-                        .and_then(|rest| rest.split('/').next())
-                        .unwrap_or("unknown");
-                    tracing::warn!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        upstream,
-                        resource_uri =
-                            crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        elapsed_ms,
-                        kind = "not_found",
-                        "upstream not connected for resource"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Failure {
-                            level: LoggingLevel::Warning,
-                            kind: "not_found",
-                        },
-                    )
-                    .await;
-                    Err(ErrorData::resource_not_found(
-                        format!("unknown resource: {uri}"),
-                        None,
-                    ))
-                }
-            };
-            return outcome;
-        }
-
-        let auth = auth_context_from_extensions(&context.extensions);
-        if let Some(oauth_subject) =
-            oauth_upstream_subject_for_request(auth, self.request_subject(&context))
-            && let Some(pool) = self.current_upstream_pool().await
-            && let Some(upstream_name) = uri
-                .strip_prefix("lab://upstream/")
-                .and_then(|rest| rest.split('/').next())
-            && let Some(config) = self.oauth_upstream_config(upstream_name).await
-        {
-            tracing::info!(
-                surface = "mcp",
-                service = "labby",
-                action = "read_resource",
-                resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                upstream = %config.name,
-                route = "subject_scoped",
-                oauth_subject = %oauth_subject,
-                "dispatch route selected"
-            );
-            let outcome = match pool
-                .subject_scoped_read_resource(&config, oauth_subject.as_ref(), uri)
-                .await
-            {
-                Ok(result) => {
-                    let elapsed_ms = start.elapsed().as_millis();
-                    tracing::info!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        subject,
-                        oauth_subject = %oauth_subject,
-                        upstream = %config.name,
-                        resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        elapsed_ms,
-                        "subject-scoped resource proxy ok"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Success,
-                    )
-                    .await;
-                    Ok(result)
-                }
-                Err(message) => {
-                    let elapsed_ms = start.elapsed().as_millis();
-                    tracing::warn!(
-                        surface = "mcp",
-                        service = "labby",
-                        action = "read_resource",
-                        upstream = %config.name,
-                        resource_uri = crate::dispatch::upstream::pool::redact_resource_uri_for_logging(uri),
-                        elapsed_ms,
-                        kind = "upstream_error",
-                        error = %message,
-                        "subject-scoped resource proxy failed"
-                    );
-                    self.emit_dispatch_notification(
-                        &context,
-                        "lab",
-                        "read_resource",
-                        elapsed_ms,
-                        DispatchLogOutcome::Failure {
-                            level: LoggingLevel::Warning,
-                            kind: "upstream_error",
-                        },
-                    )
-                    .await;
-                    Err(ErrorData::invalid_params(message, None))
-                }
-            };
-            return outcome;
-        }
-
-        let json = if uri == "lab://catalog" {
-            self.catalog_json().await
-        } else if let Some(service) = uri
-            .strip_prefix("lab://")
-            .and_then(|value| value.strip_suffix("/actions"))
-        {
-            self.service_actions_json(service).await
-        } else {
-            return Err(ErrorData::resource_not_found(
-                format!("unknown resource: {uri}"),
-                None,
-            ));
-        };
-
-        match json {
-            Ok(value) => {
-                let text = match serde_json::to_string_pretty(&value) {
-                    Ok(t) => t,
-                    Err(e) => {
-                        tracing::error!(
-                            surface = "mcp",
-                            service = "labby",
-                            action = "read_resource",
-                            subject,
-                            error = %e,
-                            "failed to serialize resource"
-                        );
-                        return Err(ErrorData::internal_error(
-                            format!("failed to serialize resource: {e}"),
-                            None,
-                        ));
-                    }
-                };
-                let elapsed_ms = start.elapsed().as_millis();
-                tracing::info!(
-                    surface = "mcp",
-                    service = "labby",
-                    action = "read_resource",
-                    subject,
-                    elapsed_ms,
-                    "resource read ok"
-                );
-                self.emit_dispatch_notification(
-                    &context,
-                    "lab",
-                    "read_resource",
-                    elapsed_ms,
-                    DispatchLogOutcome::Success,
-                )
-                .await;
-                Ok(ReadResourceResult::new(vec![
-                    ResourceContents::text(text, uri.clone()).with_mime_type("application/json"),
-                ]))
-            }
-            Err(e) => {
-                let elapsed_ms = start.elapsed().as_millis();
-                tracing::error!(
-                    surface = "mcp",
-                    service = "labby",
-                    action = "read_resource",
-                    elapsed_ms,
-                    kind = "internal_error",
-                    "resource read failed"
-                );
-                self.emit_dispatch_notification(
-                    &context,
-                    "lab",
-                    "read_resource",
-                    elapsed_ms,
-                    DispatchLogOutcome::Failure {
-                        level: LoggingLevel::Error,
-                        kind: "internal_error",
-                    },
-                )
-                .await;
-                Err(ErrorData::internal_error(e.to_string(), None))
-            }
-        }
+        self.read_resource_impl(request, context).await
     }
 
     async fn list_tools(
